Extract helm download code, refactor helm command execution
