+++ src/main/kotlin/com/deviceinsight/helm/AbstractHelmMojo.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2020 the original author or authors.
+ * Copyright 2018-2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,10 +16,17 @@
 
 package com.deviceinsight.helm
 
+import com.deviceinsight.helm.util.HelmDownloader
+import com.deviceinsight.helm.util.PlatformDetector
+import org.apache.maven.artifact.repository.ArtifactRepository
+import org.apache.maven.artifact.resolver.ArtifactResolutionRequest
 import org.apache.maven.plugin.AbstractMojo
+import org.apache.maven.plugins.annotations.Component
 import org.apache.maven.plugins.annotations.Parameter
 import org.apache.maven.project.MavenProject
+import org.apache.maven.repository.RepositorySystem
 import java.io.File
+import java.net.URI
 import kotlin.concurrent.thread
 
 abstract class AbstractHelmMojo : AbstractMojo() {
@@ -27,24 +34,47 @@ abstract class AbstractHelmMojo : AbstractMojo() {
 	@Parameter(property = "chartVersion", required = false, defaultValue = "\${project.model.version}")
 	protected lateinit var chartVersion: String
 
-	@Parameter(defaultValue = "\${project}", readonly = true, required = true)
-	protected lateinit var project: MavenProject
-
 	@Parameter(property = "chartFolder", required = false)
 	private var chartFolder: String? = null
 
-	/**
-	 * Name of the chart
-	 */
 	@Parameter(property = "chartName", required = false)
 	private var chartName: String? = null
 
-	protected fun executeCmd(
+	@Parameter(property = "helmGroupId", defaultValue = "com.deviceinsight.helm")
+	private lateinit var helmGroupId: String
+
+	@Parameter(property = "helmArtifactId", defaultValue = "helm")
+	private lateinit var helmArtifactId: String
+
+	@Parameter(property = "helmVersion", required = true, defaultValue = "3.11.2")
+	private lateinit var helmVersion: String
+
+	@Parameter(property = "helmDownloadUrl", defaultValue = "https://get.helm.sh/")
+	private lateinit var helmDownloadUrl: URI
+
+	@Parameter(readonly = true, required = true, defaultValue = "\${localRepository}")
+	private lateinit var localRepository: ArtifactRepository
+
+	@Parameter(readonly = true, required = true, defaultValue = "\${project.remoteArtifactRepositories}")
+	private lateinit var remoteRepositories: List<ArtifactRepository>
+
+	@Component
+	protected lateinit var project: MavenProject
+
+	@Component
+	private lateinit var repositorySystem: RepositorySystem
+
+	private val helmDownloader = HelmDownloader(log)
+	private lateinit var helm: String
+
+	protected fun executeHelmCmd(
 		cmd: List<String>,
 		directory: File = target(),
 		logStdoutToInfo: Boolean = false,
 		redirectOutput: ProcessBuilder.Redirect = ProcessBuilder.Redirect.PIPE
 	) {
+		initializeHelm()
+
 		check(directory.exists()) {
 			val hint = if (directory == target()) {
 				"Target helm chart dir ($directory) does not exist. Did you run the 'package' goal first?"
@@ -52,10 +82,10 @@ abstract class AbstractHelmMojo : AbstractMojo() {
 				"Working directory does not exist: $directory"
 			}
 
-			"Unable to execute '${cmd.joinToString(" ")}': $hint"
+			"Unable to execute 'helm ${cmd.joinToString(" ")}': $hint"
 		}
 
-		val proc = ProcessBuilder(cmd)
+		val proc = ProcessBuilder(listOf(helm) + cmd)
 			.directory(directory)
 			.redirectOutput(redirectOutput)
 			.redirectError(ProcessBuilder.Redirect.PIPE)
@@ -74,11 +104,59 @@ abstract class AbstractHelmMojo : AbstractMojo() {
 		stdoutPrinter.join()
 		stderrPrinter.join()
 
-		log.debug("When executing '${cmd.joinToString(" ")}' in '${directory.absolutePath}', result was ${proc.exitValue()}")
+		log.debug("When executing 'helm ${cmd.joinToString(" ")}' in '${directory.absolutePath}', result was ${proc.exitValue()}")
 
 		if (proc.exitValue() != 0) {
-			throw RuntimeException("When executing '${cmd.joinToString(" ")}' got result code '${proc.exitValue()}'")
+			throw RuntimeException("When executing 'helm ${cmd.joinToString(" ")}' got result code '${proc.exitValue()}'")
 		}
+	}
+
+	fun initializeHelm() {
+		if(!this::helm.isInitialized) {
+			checkHelmVersion()
+			resolveHelmBinary()
+		}
+	}
+
+	private fun checkHelmVersion() {
+		val versionPattern = """^(\d+)\.(\d+)\..*""".toRegex()
+		val matchResult = versionPattern.matchEntire(helmVersion)
+		requireNotNull(matchResult) { "Expected Helm version '$helmVersion' to match '$versionPattern'" }
+		val (majorVersion, minorVersion) = matchResult.destructured.toList().map(String::toInt)
+
+		if (majorVersion > 3) {
+			log.warn("This plugin was not tested with versions beyond Helm 3")
+			log.warn("Please check whether an update for this plugin is available")
+		}
+
+		require(majorVersion > 3 || (majorVersion == 3 && minorVersion >= 8)) {
+			"This plugin requires at least Helm 3.8"
+		}
+	}
+
+	private fun resolveHelmBinary() {
+		val platformIdentifier = PlatformDetector.detectHelmReleasePlatformIdentifier()
+		val helmArtifact = repositorySystem.createArtifactWithClassifier(
+			helmGroupId,
+			helmArtifactId,
+			helmVersion,
+			"binary",
+			platformIdentifier
+		)
+		val request = ArtifactResolutionRequest().apply {
+			artifact = helmArtifact
+			isResolveTransitively = false
+			localRepository = this@AbstractHelmMojo.localRepository
+			remoteRepositories = this@AbstractHelmMojo.remoteRepositories
+		}
+
+		val resolutionResult = repositorySystem.resolve(request)
+		if (!resolutionResult.isSuccess) {
+			log.info("Artifact not found in remote repositories")
+			helmDownloader.downloadAndInstallHelm(helmArtifact.file, helmVersion, helmDownloadUrl)
+		}
+
+		helm = helmArtifact.file.absolutePath
 	}
 
 
@@ -91,12 +169,4 @@ abstract class AbstractHelmMojo : AbstractMojo() {
 	protected fun chartFolder() = chartFolder ?: "src/main/helm/${chartName()}"
 
 	protected fun isChartFolderPresent() = File("${project.basedir}/${chartFolder()}").exists()
-
-	protected fun quoteFilePath(filePath: String): String =
-		if (filePath.contains(Regex("\\s"))) {
-			"\"$filePath\""
-		} else {
-			filePath
-		}
-
 }
