+++ src/main/kotlin/com/deviceinsight/helm/ResolveHelmMojo.kt
@@ -16,121 +16,14 @@
 
 package com.deviceinsight.helm
 
-import com.deviceinsight.helm.util.PlatformDetector
-import org.apache.maven.artifact.Artifact
-import org.apache.maven.artifact.repository.ArtifactRepository
-import org.apache.maven.artifact.resolver.ArtifactResolutionRequest
-import org.apache.maven.plugins.annotations.Component
 import org.apache.maven.plugins.annotations.LifecyclePhase
 import org.apache.maven.plugins.annotations.Mojo
-import org.apache.maven.plugins.annotations.Parameter
-import org.apache.maven.repository.RepositorySystem
-import java.net.URI
-import java.net.URL
-import java.nio.file.Files
-import java.nio.file.Path
-import java.nio.file.StandardCopyOption
-import java.util.zip.ZipEntry
-import java.util.zip.ZipInputStream
 
 @Mojo(name = "resolve", defaultPhase = LifecyclePhase.NONE)
 open class ResolveHelmMojo : AbstractHelmMojo() {
 
-	@Parameter(property = "helmGroupId", defaultValue = "com.deviceinsight.helm")
-	private lateinit var helmGroupId: String
-
-	@Parameter(property = "helmArtifactId", defaultValue = "helm")
-	private lateinit var helmArtifactId: String
-
-	@Parameter(property = "helmVersion", required = true, defaultValue = "3.11.2")
-	private lateinit var helmVersion: String
-
-	@Parameter(property = "helmDownloadUrl", defaultValue = "https://get.helm.sh/")
-	private lateinit var helmDownloadUrl: URI
-
-	@Parameter(readonly = true, required = true, defaultValue = "\${localRepository}")
-	private lateinit var localRepository: ArtifactRepository
-
-	@Parameter(readonly = true, required = true, defaultValue = "\${project.remoteArtifactRepositories}")
-	private lateinit var remoteRepositories: List<ArtifactRepository>
-
-	@Component
-	private lateinit var repositorySystem: RepositorySystem
-
-	protected lateinit var helm: String
-
 	override fun execute() {
-		checkHelmVersion()
-		helm = resolveHelmBinary()
+		initializeHelm()
 	}
 
-	private fun checkHelmVersion() {
-		val versionPattern = """^(\d+)\.(\d+)\..*""".toRegex()
-		val matchResult = versionPattern.matchEntire(helmVersion)
-		requireNotNull(matchResult) { "Expected Helm version '$helmVersion' to match '$versionPattern'" }
-		val (majorVersion, minorVersion) = matchResult.destructured.toList().map(String::toInt)
-
-		if (majorVersion > 3) {
-			log.warn("This plugin was not tested with versions beyond Helm 3")
-			log.warn("Please check whether an update for this plugin is available")
-		}
-
-		require(majorVersion > 3 || (majorVersion == 3 && minorVersion >= 8)) {
-			"This plugin requires at least Helm 3.8"
-		}
-	}
-
-	private fun resolveHelmBinary(): String {
-		val platformIdentifier = PlatformDetector.detectHelmReleasePlatformIdentifier()
-		val helmArtifact = repositorySystem.createArtifactWithClassifier(
-			helmGroupId,
-			helmArtifactId,
-			helmVersion,
-			"binary",
-			platformIdentifier
-		)
-		val request = ArtifactResolutionRequest().apply {
-			artifact = helmArtifact
-			isResolveTransitively = false
-			localRepository = this@ResolveHelmMojo.localRepository
-			remoteRepositories = this@ResolveHelmMojo.remoteRepositories
-		}
-
-		val resolutionResult = repositorySystem.resolve(request)
-		if (!resolutionResult.isSuccess) {
-			log.info("Artifact not found in remote repositories")
-			downloadAndInstallHelm(helmArtifact, platformIdentifier)
-		}
-
-		return helmArtifact.file.apply { setExecutable(true) }.absolutePath
-	}
-
-	private fun downloadAndInstallHelm(artifact: Artifact, platformIdentifier: String) {
-		val fileName = "helm-v$helmVersion-$platformIdentifier"
-		val url = helmDownloadUrl.resolve("./$fileName.zip").toURL()
-		val targetFile = artifact.file.toPath()
-
-		Files.createDirectories(targetFile.parent)
-		downloadFileAndExtractBinary(url, targetFile)
-	}
-
-	private fun downloadFileAndExtractBinary(url: URL, destination: Path) {
-		log.info("Downloading Helm binary from $url")
-
-		ZipInputStream(url.openStream()).use { zipInputStream ->
-			val zipEntries = generateSequence { zipInputStream.nextEntry }
-			val foundHelmBinaries = zipEntries.filter(::isHelmBinary).map {
-				Files.copy(zipInputStream, destination, StandardCopyOption.REPLACE_EXISTING)
-			}.count()
-
-			require(foundHelmBinaries == 1) {
-				"Expected 1 but found $foundHelmBinaries Helm binaries in $url"
-			}
-		}
-
-		log.info("Finished downloading Helm binary")
-	}
-
-	private fun isHelmBinary(entry: ZipEntry): Boolean =
-		!entry.isDirectory && (entry.name.endsWith("helm") || entry.name.endsWith("helm.exe"))
 }
